refactor(convert): extract block module with convert entry
Phase 1 of fulgur-odgo. Move the dispatcher's catch-all branches
(childless + container) from convert_node_inner into convert/block.rs
as `pub(super) fn convert(doc, node_id, ctx, depth) -> Box<dyn Pageable>`.

Unlike the four try_convert wrappers (Tasks 6-9) which return
`Option<Box<...>>`, block::convert is the catch-all leaf and always
returns a Pageable: childless nodes produce SpacerPageable (or a
BlockPageable when pseudos/column pagination demand a wrapper);
container nodes produce a BlockPageable wrapping
positioned::collect_positioned_children's output.

convert_node_inner's body is now: 4 try_convert if-let-Some chains
followed by a single block::convert(...) tail expression.

compute_content_box and is_non_visual_element STAY in mod.rs as
cross-module shared helpers (per residue checklist).

No behavioral change. VRT byte-identical against committed goldens.
