+++ crates/fulgur/src/convert/mod.rs
@@ -448,9 +448,6 @@ fn convert_node_inner(
     ctx: &mut ConvertContext<'_>,
     depth: usize,
 ) -> Box<dyn Pageable> {
-    let node = doc.get_node(node_id).unwrap();
-    let (width, height) = size_in_pt(node.final_layout.size);
-
     // List-item dispatch: outside marker / display:list-item fallback / inside marker — see list_item::try_convert.
     if let Some(p) = list_item::try_convert(doc, node_id, ctx, depth) {
         return p;
@@ -471,59 +468,7 @@ fn convert_node_inner(
         return p;
     }
 
-    let layout_children_guard = node.layout_children.borrow();
-    let children: &[usize] = layout_children_guard.as_deref().unwrap_or(&node.children);
-
-    if children.is_empty() {
-        let style = extract_block_style(node, ctx.assets);
-        let content_box = compute_content_box(node, &style);
-        // Check for pseudo images even on childless elements — e.g.
-        // `<div class="icon"></div>` with `.icon::before { content: url(...) }`
-        // should emit the image. Without this the pseudo is silently dropped.
-        let (positioned_children, has_pseudo) =
-            pseudo::wrap_with_pseudo_content(doc, node, ctx, depth, content_box, Vec::new());
-        let pagination = extract_pagination_from_column_css(ctx, node);
-        if style.needs_block_wrapper()
-            || has_pseudo
-            || pagination != crate::pageable::Pagination::default()
-        {
-            let (opacity, visible) = extract_opacity_visible(node);
-            let mut block = BlockPageable::with_positioned_children(positioned_children)
-                .with_pagination(pagination)
-                .with_style(style)
-                .with_opacity(opacity)
-                .with_visible(visible)
-                .with_id(extract_block_id(node));
-            block.wrap(width, height);
-            block.layout_size = Some(Size { width, height });
-            return Box::new(block);
-        }
-        // Plain leaf node — create a spacer with the computed height
-        let mut spacer = SpacerPageable::new(height);
-        spacer.wrap(width, height);
-        return Box::new(spacer);
-    }
-
-    // Container node — collect children with Taffy-computed positions
-    let positioned_children = positioned::collect_positioned_children(doc, children, ctx, depth);
-    let style = extract_block_style(node, ctx.assets);
-    let content_box = compute_content_box(node, &style);
-    let (positioned_children, _has_pseudo) =
-        pseudo::wrap_with_pseudo_content(doc, node, ctx, depth, content_box, positioned_children);
-
-    let has_style = style.needs_block_wrapper();
-    let (opacity, visible) = extract_opacity_visible(node);
-    let mut block = BlockPageable::with_positioned_children(positioned_children)
-        .with_pagination(extract_pagination_from_column_css(ctx, node))
-        .with_style(style)
-        .with_opacity(opacity)
-        .with_visible(visible)
-        .with_id(extract_block_id(node));
-    block.wrap(width, 10000.0);
-    if has_style {
-        block.layout_size = Some(Size { width, height });
-    }
-    Box::new(block)
+    block::convert(doc, node_id, ctx, depth)
 }
 
 use crate::blitz_adapter::{extract_inline_svg_tree, get_attr};
